separate tests into calendarHelpers.test.js
