+++ src/calendarHelpers.js
@@ -171,803 +171,23 @@ const rebuildAppointmentObjs = function(emptyAppointments, allAppointments, user
 
   // reassign timeslots where there are conflicts
   for (const day in sameSlotAppointments) {
-    // console.log(sameSlotAppointments[day]);
     for (const timeslot in sameSlotAppointments[day]) {
       // assign an appointment through the matchmaker
       reconstructedAppointments[day][timeslot] = matchMake(sameSlotAppointments[day][timeslot]);
     }
   }
 
-  console.log(reconstructedAppointments['WED']['00:00']);
-  // console.log(sameSlotAppointments);
   return reconstructedAppointments;
-}
-
+};
 
 // replace with static var after finalizing formats
 const allSlots = autoGenerateEmptyAppointments();
-
-// // examples
-// let today = new Date();
-// console.log(extractDayOfWeek(today.toString()));
-// console.log(extractTimeString("2021-03-29T01:07:04.353Z"));
-// console.log(changeToUserTZ(today.toString(), 'Asia/Singapore'));
-// console.log(getWeekDates("2021-03-29T19:07:04.353Z"));
-
-const fakeSessions = [
-  // 3 appointments on the same timeslot
-  {
-    id: 1,
-    owner_first_name: 'Chuck',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'd93ghjfek',
-    start_time: '2021-03-15T16:00:00.000Z',
-    activity_type: 'napping'
-  },
-  {
-    id: 2,
-    owner_first_name: 'Rick',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'dsa98f89b',
-    start_time: '2021-03-15T16:00:00.000Z',
-    activity_type: 'lounging'
-  },
-  {
-    id: 3,
-    owner_first_name: 'Morty',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'fd938hgds',
-    start_time: '2021-03-15T16:00:00.000Z',
-    activity_type: 'sleeping'
-  },
-  // 1 appointment with no dups
-  {
-    id: 4,
-    owner_first_name: 'Coolguy',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'fdsegfgr0a',
-    start_time: '2021-03-15T16:30:00.000Z',
-    activity_type: 'sleeping'
-  },
-  // 2 more dup appointments
-  {
-    id: 5,
-    owner_first_name: 'Dup1',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'hjy5eje5',
-    start_time: '2021-03-15T17:30:00.000Z',
-    activity_type: 'sleeping'
-  },
-  {
-    id: 6,
-    owner_first_name: 'Dup2',
-    owner_profile_image_url: 'avatar',
-    owner_ref_id: 'gvrf43hj6',
-    start_time: '2021-03-15T17:30:00.000Z',
-    activity_type: 'sleeping'
-  },
-];
-
-rebuildAppointmentObjs(allSlots, fakeSessions, 'Asia/Singapore');
-
 
 module.exports = {
   allSlots,
   extractTimeString,
   extractDayOfWeek,
   changeToUserTZ,
-  getWeekDates
-}
-
-/* example allSlots
-const allSlots = {
-  MON: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  TUE: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': {
-      id: null,
-      owner_first_name: null,
-      onwer_profile_image_url: null,
-      owner_ref_id: null,
-      day: 'TUE',
-      startTime: 2021-03-15T16:00:00.000Z,
-      activityType: null
-    },
-    '09:15': { state: 'empty' },
-    '09:30': {
-      id: null,
-      owner_name: null,
-      owner_pic: null,
-      day: 'TUE',
-      startTime: 2021-03-15T16:30:00.000Z,
-      activityType: null
-    },
-    '09:45': { state: 'empty' },
-    '10:00': {
-      id: null,
-      owner_name: null,
-      owner_pic: null,
-      day: 'TUE',
-      startTime: 2021-03-15T17:00:00.000Z,
-      activityType: null
-    },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  WED: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  THU: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  FRI: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  SAT: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  },
-  SUN: {
-    '00:00': { state: 'empty' },
-    '00:15': { state: 'empty' },
-    '00:30': { state: 'empty' },
-    '00:45': { state: 'empty' },
-    '01:00': { state: 'empty' },
-    '01:15': { state: 'empty' },
-    '01:30': { state: 'empty' },
-    '01:45': { state: 'empty' },
-    '02:00': { state: 'empty' },
-    '02:15': { state: 'empty' },
-    '02:30': { state: 'empty' },
-    '02:45': { state: 'empty' },
-    '03:00': { state: 'empty' },
-    '03:15': { state: 'empty' },
-    '03:30': { state: 'empty' },
-    '03:45': { state: 'empty' },
-    '04:00': { state: 'empty' },
-    '04:15': { state: 'empty' },
-    '04:30': { state: 'empty' },
-    '04:45': { state: 'empty' },
-    '05:00': { state: 'empty' },
-    '05:15': { state: 'empty' },
-    '05:30': { state: 'empty' },
-    '05:45': { state: 'empty' },
-    '06:00': { state: 'empty' },
-    '06:15': { state: 'empty' },
-    '06:30': { state: 'empty' },
-    '06:45': { state: 'empty' },
-    '07:00': { state: 'empty' },
-    '07:15': { state: 'empty' },
-    '07:30': { state: 'empty' },
-    '07:45': { state: 'empty' },
-    '08:00': { state: 'empty' },
-    '08:15': { state: 'empty' },
-    '08:30': { state: 'empty' },
-    '08:45': { state: 'empty' },
-    '09:00': { state: 'empty' },
-    '09:15': { state: 'empty' },
-    '09:30': { state: 'empty' },
-    '09:45': { state: 'empty' },
-    '10:00': { state: 'empty' },
-    '10:15': { state: 'empty' },
-    '10:30': { state: 'empty' },
-    '10:45': { state: 'empty' },
-    '11:00': { state: 'empty' },
-    '11:15': { state: 'empty' },
-    '11:30': { state: 'empty' },
-    '11:45': { state: 'empty' },
-    '12:00': { state: 'empty' },
-    '12:15': { state: 'empty' },
-    '12:30': { state: 'empty' },
-    '12:45': { state: 'empty' },
-    '13:00': { state: 'empty' },
-    '13:15': { state: 'empty' },
-    '13:30': { state: 'empty' },
-    '13:45': { state: 'empty' },
-    '14:00': { state: 'empty' },
-    '14:15': { state: 'empty' },
-    '14:30': { state: 'empty' },
-    '14:45': { state: 'empty' },
-    '15:00': { state: 'empty' },
-    '15:15': { state: 'empty' },
-    '15:30': { state: 'empty' },
-    '15:45': { state: 'empty' },
-    '16:00': { state: 'empty' },
-    '16:15': { state: 'empty' },
-    '16:30': { state: 'empty' },
-    '16:45': { state: 'empty' },
-    '17:00': { state: 'empty' },
-    '17:15': { state: 'empty' },
-    '17:30': { state: 'empty' },
-    '17:45': { state: 'empty' },
-    '18:00': { state: 'empty' },
-    '18:15': { state: 'empty' },
-    '18:30': { state: 'empty' },
-    '18:45': { state: 'empty' },
-    '19:00': { state: 'empty' },
-    '19:15': { state: 'empty' },
-    '19:30': { state: 'empty' },
-    '19:45': { state: 'empty' },
-    '20:00': { state: 'empty' },
-    '20:15': { state: 'empty' },
-    '20:30': { state: 'empty' },
-    '20:45': { state: 'empty' },
-    '21:00': { state: 'empty' },
-    '21:15': { state: 'empty' },
-    '21:30': { state: 'empty' },
-    '21:45': { state: 'empty' },
-    '22:00': { state: 'empty' },
-    '22:15': { state: 'empty' },
-    '22:30': { state: 'empty' },
-    '22:45': { state: 'empty' },
-    '23:00': { state: 'empty' },
-    '23:15': { state: 'empty' },
-    '23:30': { state: 'empty' },
-    '23:45': { state: 'empty' }
-  }
-}
-*/
+  getWeekDates,
+  rebuildAppointmentObjs
+};
